style(chatlogs): pep8
