+++ src/colors.py
@@ -123,6 +123,6 @@ class Colors:
                         f.append(int(offset * number / gradient[1] + gradients[gradient][0][rgb]))
                 return color(number, fore=f)
 
-    def escape_ansi(self, line):
+    def to_plain(self, line):
         ansi_escape = re.compile(r'(?:\x1B[@-_]|[\x80-\x9F])[0-?]*[ -/]*[@-~]')
         return ansi_escape.sub('', line)
